PMD violation fix
